+++ packages/djangoflow_websocket/lib/djangoflow_websocket.dart
@@ -1,5 +1,8 @@
 library djangoflow_websocket;
 
-export 'blocs/djangoflow_websocket_cubit/djangoflow_websocket_cubit.dart';
-export 'widgets/djangoflow_websocket_bloc_listener.dart';
-export 'widgets/djangoflow_websocket_bloc_builder.dart';
+export 'src/blocs/djangoflow_websocket_cubit/djangoflow_websocket_cubit.dart';
+export 'src/widgets/djangoflow_websocket_bloc_listener.dart';
+export 'src/widgets/djangoflow_websocket_bloc_builder.dart';
+export 'src/interfaces/djangoflow_websocket_cubit_base.dart';
+export 'src/models/djangoflow_websocket_config.dart';
+export 'src/json_parsing_exception.dart';